requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
APScheduler==3.10.4
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.18.0
